put implementations into a cpp for AbstractSubCollisionPipeline
